Fit SCORMs in Chrome
As of this writing, the latest Chrome update, Chrome 39, is not
correctly displaying embedded SCORMs. The bug has beeen noted in Chrome
and should be fixed early 2015, but in the meantime, this change seems
to solve the issue.
Oddly, setting it to 100% doesn't work.
For more information, see the Moodle tracker:
https://tracker.moodle.org/browse/MDL-48375