Excel Compatible Dates in Activity Completion Report
By default this pops out as a string that is hard to manipulate.
Because moodle timestamps count seconds and excel timestamps count days,
we divide our time by the number of seconds in a day, 86,400. Moodle
timestamps count seconds since the unix epoch (1/1/1970), but excel
counts days since 1/1/1900, so we'll also add the number of days between
these two epochs, 25,569.