Automatically Redirect to User's Sitewide Profile
When you look up a user from the Quickfind List (aka All Users List) in
a course that the user is not enrolled in, you'll simply get an error
message with the option to continue back to the course. In most cases,
what we'd rather do is go to the user's sitewide profile, so we added a
redirect.